refactor message payload code structure and naming
