+++ src/padlocal/message-parser/payload/message-emotion.ts
@@ -33,7 +33,7 @@ export interface EmojiMessagePayload {
   gameext?: string;
 }
 
-export async function emotionPayloadParser(message: PadLocal.Message.AsObject): Promise<EmojiMessagePayload> {
+export async function parseEmotionMessagePayload(message: PadLocal.Message.AsObject): Promise<EmojiMessagePayload> {
   const jsonPayload: EmotionXmlSchema = await xmlToJson(message.content);
 
   const len = parseInt(jsonPayload.msg.emoji.$.len, 10) || 0;
@@ -61,7 +61,7 @@ export async function emotionPayloadParser(message: PadLocal.Message.AsObject): 
   };
 }
 
-export function emotionPayloadGenerator(emojiMessagePayload: EmojiMessagePayload): string {
+export function generateEmotionPayload(emojiMessagePayload: EmojiMessagePayload): string {
   return `<msg><emoji cdnurl="${emojiMessagePayload.cdnurl}" len="${emojiMessagePayload.len}" md5="${
     emojiMessagePayload.md5
   }" type="${emojiMessagePayload.type}"/>${emojiMessagePayload.gameext || ""}</msg>`;
