+++ src/padlocal/message-parser/payload/message-room.ts
@@ -1,50 +1,8 @@
-/* eslint-disable camelcase */
 import type PadLocal from "padlocal-client-ts/dist/proto/padlocal_pb.js";
 import { parseSysmsgMessagePayload } from "./message-sysmsg.js";
 import type { PatMessagePayload } from "./sysmsg/message-pat.js";
 import { isContactId, isIMContactId, isIMRoomId, isRoomId } from "../../utils/is-type.js";
 import { log } from "wechaty";
-
-export interface RoomXmlSchema {
-  sysmsg: {
-    $: {
-      type: string;
-    };
-    sysmsgtemplate: {
-      content_template: {
-        $: {
-          type: string;
-        };
-        plain: string;
-        template: string;
-        link_list: {
-          link: [
-            {
-              $: {
-                name: string;
-                type: string;
-                hidden?: string;
-              };
-              memberlist?: {
-                member: [
-                  {
-                    username: string;
-                    nickname: string;
-                  }
-                ];
-              };
-              separator?: string;
-              title?: string;
-              usernamelist?: {
-                username: string;
-              };
-            }
-          ];
-        };
-      };
-    };
-  };
-}
 
 export interface RoomMessageContactInfo {
   talkerId?: string,
